feat: move extractVersionFromOutput function to version package and add tests
