+++ cluster/cluster_version.go
@@ -3,68 +3,11 @@ package cluster
 import (
 	"fmt"
 	"os/exec"
-	"regexp"
-	"strings"
 
 	"github.com/signal18/replication-manager/config"
 	"github.com/signal18/replication-manager/utils/state"
 	"github.com/signal18/replication-manager/utils/version"
 )
-
-// extractVersionFromOutput extracts the version string from command output,
-// removing paths and other noise that might confuse version parsing.
-// It looks for common version markers like "Ver", "version", "from", etc.
-func extractVersionFromOutput(output string) string {
-	// Remove leading/trailing whitespace and newlines
-	output = strings.TrimSpace(output)
-
-	// For outputs with multiple lines, take only the first line
-	if lines := strings.Split(output, "\n"); len(lines) > 0 {
-		output = lines[0]
-		output = strings.TrimSpace(output)
-	}
-
-	// Remove any file paths from the beginning (e.g., /usr/bin/mysql Ver 8.0.28...)
-	// Look for patterns like /path/to/binary and remove them
-	pathRegex := regexp.MustCompile(`^[/\w\-\.]+/[\w\-]+\s+`)
-	output = pathRegex.ReplaceAllString(output, "")
-
-	// Look for version markers and extract from that point onwards
-	// These markers indicate where the actual version information starts
-	versionMarkers := []struct {
-		marker      string
-		stripMarker bool
-	}{
-		{"Ver ", true},
-		{"version ", true},
-		{"from ", true},
-		{"v", true},
-	}
-
-	for _, m := range versionMarkers {
-		if idx := strings.Index(output, m.marker); idx != -1 {
-			// Extract from the marker onwards
-			if m.stripMarker {
-				output = output[idx+len(m.marker):]
-			} else {
-				output = output[idx:]
-			}
-			break
-		}
-	}
-
-	// Remove common tool names from the beginning if they appear without version markers
-	// (e.g., "mydumper 0.11.5" -> "0.11.5", "restic 0.15.0" -> "0.15.0")
-	toolNames := []string{"mydumper ", "restic ", "mysql ", "mysqldump ", "mysqlbinlog "}
-	for _, tool := range toolNames {
-		if strings.HasPrefix(output, tool) {
-			output = strings.TrimPrefix(output, tool)
-			break
-		}
-	}
-
-	return strings.TrimSpace(output)
-}
 
 func (cluster *Cluster) RefreshToolVersions() {
 	if err := cluster.RefreshDBClientVersion(); err != nil {
@@ -117,7 +60,7 @@ func (cluster *Cluster) RefreshDBClientVersion() error {
 
 	// Clean up the output to remove paths and extract version info
 	rawOutput := string(out)
-	vstring := extractVersionFromOutput(rawOutput)
+	vstring := version.ExtractVersionFromOutput(rawOutput)
 
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "DB client raw output: %s", rawOutput)
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "DB client cleaned version string: %s", vstring)
@@ -171,7 +114,7 @@ func (cluster *Cluster) RefreshDBClientDumpVersion() error {
 
 	// Clean up the output to remove paths and extract version info
 	rawOutput := string(out)
-	vstring := extractVersionFromOutput(rawOutput)
+	vstring := version.ExtractVersionFromOutput(rawOutput)
 
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "Dump client raw output: %s", rawOutput)
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "Dump client cleaned version string: %s", vstring)
@@ -224,7 +167,7 @@ func (cluster *Cluster) RefreshDBClientBinlogVersion() error {
 
 	// Clean up the output to remove paths and extract version info
 	rawOutput := string(out)
-	vstring := extractVersionFromOutput(rawOutput)
+	vstring := version.ExtractVersionFromOutput(rawOutput)
 
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "Binlog client raw output: %s", rawOutput)
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "Binlog client cleaned version string: %s", vstring)
@@ -277,7 +220,7 @@ func (cluster *Cluster) RefreshMyDumperVersion() error {
 
 	// Clean up the output to remove paths and extract version info
 	rawOutput := string(out)
-	vstring := extractVersionFromOutput(rawOutput)
+	vstring := version.ExtractVersionFromOutput(rawOutput)
 
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "MyDumper raw output: %s", rawOutput)
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "MyDumper cleaned version string: %s", vstring)
@@ -325,7 +268,7 @@ func (cluster *Cluster) RefreshResticVersion() error {
 
 	// Clean up the output to remove paths and extract version info
 	rawOutput := string(out)
-	vstring := extractVersionFromOutput(rawOutput)
+	vstring := version.ExtractVersionFromOutput(rawOutput)
 
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "Restic raw output: %s", rawOutput)
 	cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModGeneral, config.LvlDbg, "Restic cleaned version string: %s", vstring)
